feat: prototype for traffic visualisation
Refs: XRDDEV-2894
