+++ src/security-server/admin-service/ui/src/views/Diagnostics/Traffic/TrafficView.vue
@@ -106,8 +106,8 @@
       </v-card-text>
     </v-card>
 
-    <v-card>
-      <TrafficChart></TrafficChart>
+    <v-card height="20rem">
+      <TrafficChart :series="series" />
     </v-card>
   </XrdTitledView>
 </template>
@@ -137,4 +137,28 @@ const dateAdapter = useDate();
 function formatDate(date: Date): string {
   return dateAdapter.toISO(date);
 }
+
+function generateRandomData(min = 0, max = 100) {
+  const data = [];
+  const now = new Date().getTime();
+  let date = new Date(now - 7 * 24 * 60 * 60 * 1000).getTime();
+  do {
+    date = new Date(date + 60 * 60 * 1000).getTime();
+    data.push([date, min + Math.floor(Math.random() * (max - min))]);
+  } while (date < now);
+  return data;
+}
+
+const series = [
+  {
+    name: 'Successful requests',
+    color: '#00d54d',
+    data: generateRandomData(40, 100),
+  },
+  {
+    name: 'Failed requests',
+    color: '#d84957',
+    data: generateRandomData(0, 5),
+  },
+];
 </script>
